requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
deepgram-sdk==3.0.3
websockets==12.0
python-multipart==0.0.6
openai==1.3.5
verboselogs==1.7 